requirements.txt
fastapi
uvicorn
mediapipe
opencv-python
websockets
aiofiles
python-multipart
sqlmodel
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
requests